+++ packages/waggle/src/libp2p/connectionsManager.test.ts
@@ -107,4 +107,89 @@ describe('Connections manager', () => {
     await expect(connectionsManager.sendCertificateRegistrationRequest('http://invalid.onion', 'cert', retryCount)).rejects.toThrow()
     expect(spyOnFetchRetry).toHaveBeenCalledTimes(retryCount)
   })
+
+  it('inits libp2p', async () => {
+    const peerId = await PeerId.create()
+    const port = 1234
+    const address = '0.0.0.0'
+    const ports = await utils.getPorts()
+    connectionsManager = new ConnectionsManager({
+      agentHost: 'localhost',
+      agentPort: ports.socksPort,
+      httpTunnelPort: ports.httpTunnelPort,
+      // @ts-expect-error
+      io: new utils.DummyIOServer(),
+      options: {
+        env: {
+          appDataPath: tmpAppDataPath
+        },
+        torControlPort: ports.controlPort
+      }
+    })
+    const localAddress = connectionsManager.createLibp2pAddress(address, port, peerId.toB58String())
+    const listenAddress = connectionsManager.createLibp2pListenAddress(address, port)
+    const result = await connectionsManager.initLibp2p(
+      peerId,
+      address,
+      port,
+      ['some/address'],
+      null,
+      port
+    )
+    expect(result.localAddress).toBe(localAddress)
+    expect(result.libp2p.peerId).toBe(peerId)
+    console.log(result.libp2p.addresses.listen)
+    expect(result.libp2p.addresses.listen).toStrictEqual([listenAddress])
+  })
+
+  it.each([
+    ['ws'],
+    ['wss']
+  ])('creates libp2p address with proper ws type (%s)', async (wsType: 'ws' | 'wss') => {
+    const address = '0.0.0.0'
+    const port = 1234
+    const peerId = await PeerId.create()
+    const ports = await utils.getPorts()
+    connectionsManager = new ConnectionsManager({
+      agentHost: 'localhost',
+      agentPort: ports.socksPort,
+      httpTunnelPort: ports.httpTunnelPort,
+      // @ts-expect-error
+      io: new utils.DummyIOServer(),
+      options: {
+        env: {
+          appDataPath: tmpAppDataPath
+        },
+        torControlPort: ports.controlPort,
+        wsType
+      }
+    })
+    const libp2pAddress = connectionsManager.createLibp2pAddress(address, port, peerId.toB58String())
+    expect(libp2pAddress).toStrictEqual(`/dns4/${address}/tcp/${port}/${wsType}/p2p/${peerId.toB58String()}`)
+  })
+
+  it.each([
+    ['ws'],
+    ['wss']
+  ])('creates libp2p listen address with proper ws type (%s)', async (wsType: 'ws' | 'wss') => {
+    const address = '0.0.0.0'
+    const port = 1234
+    const ports = await utils.getPorts()
+    connectionsManager = new ConnectionsManager({
+      agentHost: 'localhost',
+      agentPort: ports.socksPort,
+      httpTunnelPort: ports.httpTunnelPort,
+      // @ts-expect-error
+      io: new utils.DummyIOServer(),
+      options: {
+        env: {
+          appDataPath: tmpAppDataPath
+        },
+        torControlPort: ports.controlPort,
+        wsType
+      }
+    })
+    const libp2pListenAddress = connectionsManager.createLibp2pListenAddress(address, port)
+    expect(libp2pListenAddress).toStrictEqual(`/dns4/${address}/tcp/${port}/${wsType}`)
+  })
 })